EPHEMERIS: Use OS tzdata for leap second calculation.
The tzdata database, also known as the Olsen, database seems to be the
definitive source for leap second data.  It is maintained regularly by folks
who follow this stuff.  The suggestion to use this database came from:
https://stackoverflow.com/questions/19332902/extract-historic-leap-seconds-from-tzdata

An alterate approach would be to include the tz repository as a submodule but
this would quickly go stale.  Therefore, I opted to pull the file from tzdata
which should be regularly updated by the OS.  The current location is suitable
for Ubuntu, not sure about other OSes.